+++ src/js/color_change_event.js
@@ -1,32 +1,37 @@
-/*
+/**
  * Custom Color Change Event
  */
 
 // Custom color change event function
-function colorChange(color, elem) {
-	console.log(color);
+colorPickerComp.colorChange = function (color, elem) {
 	// Defining the RGBA value conversion
-	let rgbaValue = HSLAToRGBA(color.h, color.s, color.l, color.a);
+	const rgbaValue = colorPickerComp.HSLAToRGBA(color.h, color.s, color.l, color.a);
+	const hex = colorPickerComp.HSLAToRGBA(color.h, color.s, color.l, color.a, true);
 
 	// Creating the event
 	const event = new CustomEvent('colorChange', {
 		// Adding the response details
 		detail: {
 			color: {
+				hsl: `hsla(${color.h}, ${color.s}%, ${color.l}%)`,
+				rgb: `rgba(${rgbaValue.r}, ${rgbaValue.g}, ${rgbaValue.b})`,
+				hex: hex,
 				hsla: `hsla(${color.h}, ${color.s}%, ${color.l}%, ${color.a})`,
 				rgba: `rgba(${rgbaValue.r}, ${rgbaValue.g}, ${rgbaValue.b}, ${rgbaValue.a})`,
-				hexa: HSLAToRGBA(color.h, color.s, color.l, color.a, true)
+				hexa: hex
 			}
 		}
 	});
 
 	// Defining element
-	const element = elem === undefined ? colorPicker.instance.element : elem;
+	const element = elem === undefined ? colorPickerComp.instance.element : elem;
+
+	// Defining color
 
 	// Changing color attributes
-	element.setAttribute('data-color', color);
-	element.style.background = color;
+	element.setAttribute('data-color', hex);
+	element.style.background = hex;
 
 	// Dispatching the event for the active object
 	element.dispatchEvent(event);
-}
+};