+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 ---
 
+### 1.2.0
+
+-   Add function to get an array of saved colors
+
+-   Made functions belong to an object instead of them being global
+
+-   Added animation when opening the color picker
+
+-   Fixed issues that arise when color is not set when the color picker instance is initiated
+
+### 1.1.0
+
+-   Added multiple output formats to select from when you reference the event info, inside the listener for the color picker.
+
 ### 1.0.1
 
 -   Put a light grey border around custom color previews so that lighter colors, like white, can be seen